requirements.txt
annotated-types
anyio
appdirs
beautifulsoup4
boto3
certifi
charset-normalizer
click
colorama
cython
dnspython
fastapi
frozendict
greenlet
h11
html5lib
httptools
hypercorn
idna
lxml
matplotlib
multitasking
numpy
pandas
peewee
psycopg2
pydantic
pydantic_core
pymongo
python-dateutil
python-dotenv
pytz
requests
scipy
setuptools
six
sniffio
soupsieve
SQLAlchemy
starlette
typing_extensions
tzdata
urllib3
uvicorn
uvloop
webencodings
wheel
yfinance